moving useful logic out of app-clipper package
