+++ packages/app-clipper/package.json
@@ -8,5 +8,8 @@
   "description": "Joplin Web Clipper",
   "main": "main.js",
   "author": "Laurent Cozic",
-  "license": "AGPL-3.0-or-later"
+  "license": "AGPL-3.0-or-later",
+  "dependencies": {
+    "@joplin/lib": "~2.13"
+  }
 }
